+++ cosmwasm/enclaves/shared/contract-engine/src/external/ecalls.rs
@@ -53,7 +53,7 @@ unsafe fn ecall_allocate_impl(buffer: *const u8, length: usize) -> EnclaveBuffer
         return EnclaveBuffer::default();
     }
 
-    validate_const_ptr!(buffer, length as usize, EnclaveBuffer::default());
+    validate_const_ptr!(buffer, length, EnclaveBuffer::default());
 
     let slice = std::slice::from_raw_parts(buffer, length);
     let result = panic::catch_unwind(|| {
@@ -181,10 +181,10 @@ pub unsafe extern "C" fn ecall_init(
 
     let failed_call = || result_init_success_to_initresult(Err(EnclaveError::FailedFunctionCall));
     validate_mut_ptr!(used_gas as _, std::mem::size_of::<u64>(), failed_call());
-    validate_const_ptr!(env, env_len as usize, failed_call());
-    validate_const_ptr!(msg, msg_len as usize, failed_call());
-    validate_const_ptr!(contract, contract_len as usize, failed_call());
-    validate_const_ptr!(sig_info, sig_info_len as usize, failed_call());
+    validate_const_ptr!(env, env_len, failed_call());
+    validate_const_ptr!(msg, msg_len, failed_call());
+    validate_const_ptr!(contract, contract_len, failed_call());
+    validate_const_ptr!(sig_info, sig_info_len, failed_call());
 
     let contract = std::slice::from_raw_parts(contract, contract_len);
     let env = std::slice::from_raw_parts(env, env_len);
@@ -267,10 +267,10 @@ pub unsafe extern "C" fn ecall_handle(
     let failed_call =
         || result_handle_success_to_handleresult(Err(EnclaveError::FailedFunctionCall));
     validate_mut_ptr!(used_gas as _, std::mem::size_of::<u64>(), failed_call());
-    validate_const_ptr!(env, env_len as usize, failed_call());
-    validate_const_ptr!(msg, msg_len as usize, failed_call());
-    validate_const_ptr!(contract, contract_len as usize, failed_call());
-    validate_const_ptr!(sig_info, sig_info_len as usize, failed_call());
+    validate_const_ptr!(env, env_len, failed_call());
+    validate_const_ptr!(msg, msg_len, failed_call());
+    validate_const_ptr!(contract, contract_len, failed_call());
+    validate_const_ptr!(sig_info, sig_info_len, failed_call());
 
     let contract = std::slice::from_raw_parts(contract, contract_len);
     let env = std::slice::from_raw_parts(env, env_len);
@@ -406,9 +406,9 @@ unsafe fn ecall_query_impl(
 
     let failed_call = || result_query_success_to_queryresult(Err(EnclaveError::FailedFunctionCall));
     validate_mut_ptr!(used_gas as _, std::mem::size_of::<u64>(), failed_call());
-    validate_const_ptr!(env, env_len as usize, failed_call());
-    validate_const_ptr!(msg, msg_len as usize, failed_call());
-    validate_const_ptr!(contract, contract_len as usize, failed_call());
+    validate_const_ptr!(env, env_len, failed_call());
+    validate_const_ptr!(msg, msg_len, failed_call());
+    validate_const_ptr!(contract, contract_len, failed_call());
 
     let contract = std::slice::from_raw_parts(contract, contract_len);
     let env = std::slice::from_raw_parts(env, env_len);
