+++ cosmwasm/enclaves/shared/contract-engine/src/wasm3/mod.rs
@@ -625,7 +625,7 @@ impl<'m> CWMemory<'m> {
         }
 
         let data_len = self.get_u32_at(region_ptr + (SIZE_OF_U32 as u32) * 2)? as usize;
-        let mut remaining_len = data_len as usize;
+        let mut remaining_len = data_len;
 
         let data = self.memory.as_slice().get(data_ptr..data_ptr + data_len);
         let data = data.ok_or(WasmEngineError::MemoryReadError)?;
